requirements.txt
flask
gunicorn
pillow
numpy==2.2.6
opencv-python==4.12.0.88
